% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_rapid_helpers.R
\name{combine_matrix}
\alias{combine_matrix}
\title{combine_matrix}
\usage{
combine_matrix(data)
}
\arguments{
\item{data}{data from matrix_counts_vec() and matrix_cells_vec()}
}
\description{
combine_matrix
}
