% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_rapid.R
\name{get_table_rapid}
\alias{get_table_rapid}
\title{Stratify Person Table Rapidly}
\usage{
get_table_rapid(persondf, strata = c(), break_yr = 5)
}
\arguments{
\item{persondf}{data.frame like object containing one row per person with the required demographic information}

\item{strata}{character vector identify additional variables (found in persondf) to stratify
person-time. By default, gender and race are included.}

\item{break_yr}{atomic numeric identifying the number of years to stratify age and calendar periods.}
}
\value{
A data.frame with a row for each strata containing the number of observed
deaths within each of the defined codes and the number of person days (pdays).
}
\description{
\code{get_table_rapid} reads in a data.frame/tibble containing basic demographic information
for each person of the cohort and stratifies the person-time and deaths into age,
calendar period, race, and sex strata. See \code{Details} for information on how the
person file must be formatted.
}
\details{
This method uses the Maxaluso's method B which makes use of the regular intervals of age and
calendar-period to speed up computation time.

The persondf tibble must contain the variables:
\itemize{
\item id,
\item gender (character: 'M'/'F'),
\item race (character: 'W'/'N'),
\item dob (date),
\item pybegin (date),
\item dlo	(date),
\item code (character: identifying outcomes of interest)
}
}
\examples{
library(LTASR)
library(dplyr)

Import example person file
person <- person_example \%>\%
    mutate(dob = as.Date(dob, format='\%m/\%d/\%Y'),
    pybegin = as.Date(pybegin, format='\%m/\%d/\%Y'),
    dlo = as.Date(dlo, format='\%m/\%d/\%Y'))

Stratify person table
py_table <- get_table_rapid(person, strata = stratify = c("gender", "race"))
}
