% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_rapid_helpers.R
\name{list_count_helper}
\alias{list_count_helper}
\title{list_count_helper}
\usage{
list_count_helper(
  s1,
  s2,
  l1,
  l2,
  first,
  odd,
  even,
  last,
  even_last,
  count0start,
  count0starteven,
  count0lasteven,
  count0lastodd,
  ov,
  ev,
  start1,
  stop1
)
}
\arguments{
\item{s1}{start col}

\item{s2}{start row}

\item{l1}{end col}

\item{l2}{end row}

\item{first}{first cell events}

\item{odd}{odds cell counts events}

\item{even}{even cell counts events}

\item{last}{last cell events}

\item{even_last}{bool, is the last cell even}

\item{count0start}{starting col or row}

\item{count0starteven}{offset if starting count cell is even}

\item{count0lasteven}{offset if last count cell is even}

\item{count0lastodd}{offset if last count cell is odd}

\item{ov}{offset to determine the cell count order}

\item{ev}{offset if starting from the first cell}

\item{start1}{first year of follow-up}

\item{stop1}{last year of follow-up}
}
\description{
list_count_helper
}
