% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_rapid_helpers.R
\name{matrix_counts_vec}
\alias{matrix_counts_vec}
\title{matrix_counts_vec}
\usage{
matrix_counts_vec(data, time_break, matrix_obj)
}
\arguments{
\item{data}{data from get_table_rapid()}

\item{time_break}{number of years defined by user}

\item{matrix_obj}{data.frame of total observed events. Arranged by years (columns) and age (rows)}
}
\description{
matrix_counts_vec
}
