% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_rapid_helpers.R
\name{matrix_helper}
\alias{matrix_helper}
\title{matrix_helper}
\usage{
matrix_helper(eos1, eos2, oos1, oos2, even_list, odd_list, matrix_obj)
}
\arguments{
\item{eos1}{offset cols, odd data}

\item{eos2}{offset rows, even data}

\item{oos1}{offset cols, odd data}

\item{oos2}{offset rows, odd data}

\item{even_list}{even cells of person-years or events count}

\item{odd_list}{odd cells of person-years or events count}

\item{matrix_obj}{person-years or events count data.frame. Arranged by years (columns) and age (rows)}
}
\description{
matrix_helper
}
