% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_table_annual_exposure.R
\name{py_table_annual_exposure}
\alias{py_table_annual_exposure}
\title{Stratify Person Table With Time Varying Annual Exposure}
\usage{
py_table_annual_exposure(
  person_data,
  exposure_data,
  outcomes,
  cat_vars,
  cont_vars,
  exp_vars,
  bin = 500,
  mid_year = 0.4516
)
}
\arguments{
\item{person_data}{data.frame like object containing one row per person with the required demographic information}

\item{exposure_data}{data.frame like object containing one row per person per year. See
Details for required variables.}

\item{outcomes}{character vector identifying the outcome variables (found in persondf)
where 1 represents a case and 0 represents no case.}

\item{cat_vars}{character vector identifying the categorical variables in person_data on which to stratify}

\item{cont_vars}{a list containing continuous variables in person_data on which to stratify,
modeled as exp_strata objects created by exp_strata()}

\item{exp_vars}{a list containing continuous variables in exposure_data on which to stratify,
modeled as exp_strata objects created by exp_strata()}

\item{bin}{a number specifying how many persons to stratify at a time. Default is 500.}

\item{mid_year}{a number specifying the proportion of the year to define as the middle when applying a lag. Default is 0.4516}
}
\value{
A data.frame with a row for each strata containing the number of observed
deaths within each of the defined codes and the number of person days (pdays).
}
\description{
\code{py_table_annual_exposure} reads in a data.frame/tibble containing basic demographic information
for each person of the cohort and stratifies the person-time and deaths into age,
calendar period, race, and sex strata. See \code{Details} for information on how the
person file must be formatted.
}
\details{
The person_data tibble must contain the (case specific) variables:
\itemize{
\item id,
\item dob (date),
\item pybegin (date),
\item dlo	(date),
\item hire_date	(date)
\item term_date (date)
The exposure_data tibble must contain the (case specific) variables:
\item id,
\item year (date),
\item exposure_variable_name (numeric),
}
}
\examples{
library(LTASR)
library(tidyr)
library(powerjoin)
library(lubridate)

outcomes <- c("LUNG", "CIRCULATORY", "CVD","RESPIRATORY", "CANCER")
cat_vars <- c("SEX")
exp1 <- exp_strata(var = 'ds',
  cutpt =  c(0,1e-10, 5,10,20,30,40,50,60,70,80,90,100,125,150,175,200,250,
    300,500,750,1000,1250,1500,2000,3000,4000,5000,6000,Inf),
  lag = 5) # The second cutpt value should be minimally above zero if you wish for no exposure to be in a separate group
exp_vars <- list(exp1)

cont1 <- exp_strata(var = 'age',
  cutpt = c(-Inf, seq(20,90, 5), Inf),
   lag = 0)
cont2 <- exp_strata(var = 'year',
  cutpt = c(seq(1950,1995,5), Inf),
  lag = 0)
cont3 <- exp_strata(var = 'ndur',
  cutpt = c(0,0.5,1,5,10,20,30, Inf),
  lag = 0)
cont_vars <- list(cont1, cont2,cont3)

py_table_annual_exposure(person_data = person_data_example,
                       exposure_data = exposure_data_example,
                       outcomes = outcomes,
                       cat_vars = cat_vars,
                       cont_vars = cont_vars,
                       exp_vars = exp_vars)
}
