% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_le.R
\name{theme_le}
\alias{theme_le}
\title{Louis' preferred ggplot2 theme}
\usage{
theme_le(...)
}
\arguments{
\item{...}{see \code{theme()}}
}
\description{
\code{theme_le} may be used as a \code{theme()} function. The theme is a slightly modified version of \code{theme_classic}.
Changes included an increased default font size and default serif font.
}
